- Support for max 4x ADS1115 (OSPi) 0x48 / 0x49 / 0x4A / 0x4B - fixt compile error (OSPi)

+++ utils.cpp
@@ -639,7 +639,7 @@ size_t freeMemory() {
 }
 #else
 
-long freeMemory() {
+size_t freeMemory() {
 	return -1;
 }
 #endif
